Transparent background color css rules
It sets the background to transparent on the items where it is set. Also removes the borders of the help boxes.
Sets the background color of the inputs to transparent. On case of focus state, it gets the white background
color back. The text input color are black of white, depends on the background color (or state).